Move APIManager to base project
