+++ PerformanceCalculator/ProcessorWorkingBeatmap.cs
@@ -66,7 +66,7 @@ namespace PerformanceCalculator
             if (!File.Exists(cachePath))
             {
                 Console.WriteLine($"Downloading {beatmapId}.osu...");
-                new FileWebRequest(cachePath, $"{Program.ENDPOINT_CONFIGURATION.WebsiteRootUrl}/osu/{beatmapId}").Perform();
+                new FileWebRequest(cachePath, $"{APIManager.ENDPOINT_CONFIGURATION.WebsiteRootUrl}/osu/{beatmapId}").Perform();
             }
 
             return new ProcessorWorkingBeatmap(cachePath, beatmapId);
