+++ PerformanceCalculator/Program.cs
@@ -6,7 +6,6 @@ using System.Text;
 using McMaster.Extensions.CommandLineUtils;
 using osu.Framework.Logging;
 using osu.Game.Beatmaps.Formats;
-using osu.Game.Online;
 using PerformanceCalculator.Difficulty;
 using PerformanceCalculator.Leaderboard;
 using PerformanceCalculator.Performance;
@@ -25,8 +24,6 @@ namespace PerformanceCalculator
     [HelpOption("-?|-h|--help")]
     public class Program
     {
-        public static readonly EndpointConfiguration ENDPOINT_CONFIGURATION = new ProductionEndpointConfiguration();
-
         public static void Main(string[] args)
         {
             Console.OutputEncoding = Encoding.UTF8;
