+++ PerformanceCalculatorGUI/PerformanceCalculatorGame.cs
@@ -13,8 +13,9 @@ using osu.Game.Graphics.Cursor;
 using osu.Game.Overlays;
 using osu.Game.Rulesets.Mods;
 using osu.Game.Rulesets.Osu;
+using PerformanceCalculator;
+using PerformanceCalculator.Configuration;
 using PerformanceCalculatorGUI.Components;
-using PerformanceCalculatorGUI.Configuration;
 
 namespace PerformanceCalculatorGUI
 {
@@ -26,7 +27,7 @@ namespace PerformanceCalculatorGUI
         // This overwrites OsuGameBase's SelectedMods to make sure it can't tweak mods when we don't want it to
         [Cached]
         [Cached(typeof(IBindable<IReadOnlyList<Mod>>))]
-        private readonly Bindable<IReadOnlyList<Mod>> mods = new(Array.Empty<Mod>());
+        private readonly Bindable<IReadOnlyList<Mod>> mods = new Bindable<IReadOnlyList<Mod>>(Array.Empty<Mod>());
 
         [Resolved]
         private FrameworkConfigManager frameworkConfig { get; set; }
