+++ PerformanceCalculatorGUI/Screens/LeaderboardScreen.cs
@@ -20,9 +20,9 @@ using osu.Game.Rulesets;
 using osu.Game.Rulesets.Mods;
 using osu.Game.Users;
 using PerformanceCalculator;
+using PerformanceCalculator.Configuration;
 using PerformanceCalculatorGUI.Components;
 using PerformanceCalculatorGUI.Components.TextBoxes;
-using PerformanceCalculatorGUI.Configuration;
 
 namespace PerformanceCalculatorGUI.Screens
 {
